requirements.txt
opencv-python-headless>=4.5.0
numpy>=1.19.0
PyYAML>=5.4.0
# rknn-toolkit-lite2>=2.3.0
ultralytics>=8.3.220
websockets>=12.0
python-dotenv>=1.0.0
modbus_tk

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0